Move currency alert features into the currencies module
